pyproject.toml
[tool.poetry]
name = "xai-budgeting"
version = "0.1.0"
description = ""
authors = ["Marc Grau <marcchristopher.grau@unisg.ch>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pdoc = "^14.3.0"
pre-commit = "^3.6.0"
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
interrogate = "^1.5.0"
requests = "^2.31.0"
mypy = "^1.8.0"
types-requests = "^2.31.0.20240106"
pandas = "2.1.4"
openpyxl = "^3.1.2"
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
catboost = "^1.2.2"
scikit-learn = "^1.4.0"
optuna = "^3.5.0"
numpy = "^1.26.3"
pyarrow = "10.0.1"
xgboost = "^2.0.3"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"
taipy = "^3.0.0"
darts = "^0.29.0"
prophet = "^1.1.5"
tensorflow = "^2.17.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
